requirements.txt
torch
transformers>=4.40.0
peft>=0.10.0
bitsandbytes
accelerate
datasets
Pillow
pyyaml
numpy
pandas